+++ Sources/Payments/Beneficiary.cs
@@ -5,7 +5,7 @@ using System.Text.Json.Serialization;
 /// <summary>
 /// Represents a beneficiary.
 /// </summary>
-public class Beneficiary: IEquatable<Beneficiary> {
+public class Beneficiary {
 
 	/// <summary>
 	/// The bank account of the beneficiary.
@@ -47,49 +47,4 @@ public class Beneficiary: IEquatable<Beneficiary> {
 	/// </summary>
 	[JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
 	public ValidationStatus? ValidationStatus { get; set; }
-
-	/// <summary>
-	/// Determines whether the two specified objects are equal.
-	/// </summary>
-	/// <param name="object1">The first object.</param>
-	/// <param name="object2">The second object.</param>
-	/// <returns><see langword="true"/> if <paramref name="object1"/> equals <paramref name="object2"/>, otherwise <see langword="false"/>.</returns>
-	public static bool operator ==(Beneficiary? object1, Beneficiary? object2) =>
-		ReferenceEquals(object1, object2) || (object1?.Equals(object2) ?? false);
-
-	/// <summary>
-	/// Determines whether the two specified objects are not equal.
-	/// </summary>
-	/// <param name="object1">The first object.</param>
-	/// <param name="object2">The second object.</param>
-	/// <returns><see langword="true"/> if <paramref name="object1"/> does not equal <paramref name="object2"/>, otherwise <see langword="false"/>.</returns>
-	public static bool operator !=(Beneficiary? object1, Beneficiary? object2) => !(object1 == object2);
-
-	/// <summary>
-	/// Determines whether the specified object is equal to this object.
-	/// </summary>
-	/// <param name="other">An object to compare with this object.</param>
-	/// <returns><see langword="true"/> if the specified object is equal to this object, otherwise <see langword="false"/>.</returns>
-	public override bool Equals(object? other) => Equals(other as Beneficiary);
-
-	/// <summary>
-	/// Determines whether the specified object is equal to this object.
-	/// </summary>
-	/// <param name="other">An object to compare with this object.</param>
-	/// <returns><see langword="true"/> if the specified object is equal to this object, otherwise <see langword="false"/>.</returns>
-	public bool Equals(Beneficiary? other) => other is not null &&
-		BankAccount == other.BankAccount &&
-		CompanyLegalIdentifier == other.CompanyLegalIdentifier &&
-		Id == other.Id &&
-		Name == other.Name &&
-		PostalAddress == other.PostalAddress &&
-		UncertaintyStatus == other.UncertaintyStatus &&
-		ValidationStatus == other.ValidationStatus;
-
-	/// <summary>
-	/// Gets the hash code for this object.
-	/// </summary>
-	/// <returns>The hash code for this object.</returns>
-	public override int GetHashCode() =>
-		HashCode.Combine(BankAccount, CompanyLegalIdentifier, Id, Name, PostalAddress, UncertaintyStatus, ValidationStatus);
 }