+++ ChangeLog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Version [0.6.0](https://github.com/MC2IT/Agicap.net/compare/v0.5.0...v0.6.0)
-- Added the `SynchronizedBeneficiary.BankAccount` property.
+- Added the `Payments.SynchronizedBeneficiary.BankAccount` property.
+- Removed the `IEquatable` implementation from the `BankAccount`, `Beneficiary` and `PostalAddress` classes of the `Payments` namespace.
 
 ## Version [0.5.0](https://github.com/MC2IT/Agicap.net/compare/v0.4.0...v0.5.0)
 - Added the `NestedList` class.